+++ src/app/auth/auth.service.ts
@@ -12,6 +12,7 @@ export class AuthService {
   private isAuthenticated = false;
   private token: string;
   private tokenTimer: any;
+  private userId: string;
   private authStatusListener = new Subject<boolean>()
 
   constructor(private http: HttpClient, private router: Router){}
@@ -22,6 +23,10 @@ export class AuthService {
 
   getIsAuth() {
     return this.isAuthenticated;
+  }
+
+  getUserid() {
+    return this.userId;
   }
 
   getAuthStatusListener() {
@@ -38,7 +43,11 @@ export class AuthService {
 
   login(email: string, password: string) {
     const authData: AuthData = {email: email, password: password};
-    this.http.post<{ token: string, expiresIn: number }>("http://localhost:3000/api/user/login", authData)
+    this.http
+    .post<{ token: string, expiresIn: number, userId: string }>(
+      "http://localhost:3000/api/user/login",
+      authData
+    )
     .subscribe(response =>{
       const token = response.token;
       this.token = token;
@@ -46,10 +55,11 @@ export class AuthService {
         const expiresInDuration = response.expiresIn;
         this.setAuthTimer(expiresInDuration);
         this.isAuthenticated = true;
+        this.userId = response.userId;
         this.authStatusListener.next(true);
         const now = new Date();
         const expirationDate = new Date(now.getTime() + expiresInDuration * 1000);
-        this.saveAuthData(token, expirationDate);
+        this.saveAuthData(token, expirationDate, this.userId);
         this.router.navigate(['/']);
       }
     });
@@ -65,6 +75,7 @@ export class AuthService {
     if(expiresIn > 0) {
       this.token = authInformation.token;
       this.isAuthenticated = true;
+      this.userId = authInformation.userId;
       this.setAuthTimer(expiresIn / 1000);
       this.authStatusListener.next(true);
     }
@@ -75,6 +86,7 @@ export class AuthService {
     this.isAuthenticated = false;
     this.authStatusListener.next(false);
     this.router.navigate(['/']);
+    this.userId = null;
     this.clearAuthData();
     clearTimeout(this.tokenTimer);
   }
@@ -85,26 +97,30 @@ export class AuthService {
     }, duration * 1000);
   }
 
-  private saveAuthData(token: string, expirationDate: Date) {
+  private saveAuthData(token: string, expirationDate: Date, userId: string) {
     localStorage.setItem("token", token);
     localStorage.setItem("expiration", expirationDate.toISOString());
+    localStorage.setItem("userId", userId);
   }
 
   private clearAuthData() {
     localStorage.removeItem("token");
     localStorage.removeItem("expiration");
+    localStorage.removeItem("userId");
   }
 
   private getAuthData() {
     const token = localStorage.getItem("token");
     const expirationDate = localStorage.getItem("expiration");
+    const userId = localStorage.getItem("userId");
     if (!token || !expirationDate) {
       return;
     }
     else {
       return {
         token: token,
-        expirationDate: new Date(expirationDate)
+        expirationDate: new Date(expirationDate),
+        userId: userId
       }
     }
   }